src/include/gpio.h: Add the mising uint8 cast

+++ src/include/gpio.h
@@ -1,10 +1,12 @@
 #ifndef GPIO_H
 #define GPIO_H
 
+#include <stdint.h>
+
 void setup_gpio();
 
 #define CAST_PINS_TO_UINT8(p0, p1, p2, p3, p4, p5, p6, p7)  \
-    (p0 | \
+    (uint8_t)(p0 | \
     (p1 << 1) | \
     (p2 << 2) | \
     (p3 << 3) | \
